Change LocalToGlobal/GlobalToLocal prototypes
Use `uint32_t` rather than `void *` for global addresses - leverages
type safety and reduces the amount of casting required, as hardware
registers usually want uint32_t anyway.

+++ source/hal/source/components/npu/ethosu_npu_init.c
@@ -121,5 +121,5 @@ int arm_ethosu_npu_init(void)
 
 uint64_t ethosu_base_pointer_remap(const void *p)
 {
-    return (uint64_t)LocalToGlobal(p);
+    return LocalToGlobal(p);
 }
